+++ ur_robot_driver/launch/ur5e.launch.py
@@ -70,6 +70,7 @@ def generate_launch_description():
                 "forward_velocity_controller",
                 "forward_position_controller",
                 "freedrive_mode_controller",
+                "passthrough_trajectory_controller",
             ],
         )
     )